SCI: Widen the font interface from uint16 to uint32 characters
Stage 1 of docs/i18n/M8_CODEPOINT_PLAN.md: a type change with no behavioural
change, so the rest of the move to code points has somewhere to land.

GfxFont's five methods - isDoubleByte, getCharWidth, getCharHeight, draw,
drawToBuffer - and all four implementations now take uint32. The value they
receive is still exactly what it was: a packed byte pair, lead byte in the
low half, assembled by GfxText16/GfxText32 and decoded by each font. Nothing
reinterprets it. 39 insertions, 39 deletions, all of them the type.

uint16 could not hold a code point above U+FFFF, so this is a precondition
for the later stages rather than an improvement on its own.

Verified by glyph sequence rather than by screenshot. A probe logged every
(character, top, left) triple passed to GfxFontFromResource::draw and
GfxFontKorean::draw across a full KQ1 play session, before and after the
change:

  Korean via SCITRS: 1076 glyphs, sequence byte-for-byte IDENTICAL
  English:           1412 vs 1425

The English difference is harness non-determinism, not a regression: re-running
the SAME post-change build produced 1412, alternating with 1425. Screenshot
diffing was tried first and abandoned for the same reason - two runs of one
unchanged build differ by up to 7722 pixels because character animation phase
varies, so pixel equality cannot decide this question. The Korean sequence is
the real evidence: it exercises the double-byte path, and it did not move.

Also built and tested with ENABLE_SCI32 on, which compiles text32.cpp's three
byte-pair packing sites and the SCI32 font paths that the default
configuration leaves out: 0 errors, 409/409 tests, both ways.

+++ engines/sci/graphics/fontkorean.cpp
@@ -47,7 +47,7 @@ GuiResourceId GfxFontKorean::getResourceId() {
 }
 
 // Returns true for first byte of double byte characters
-bool GfxFontKorean::isDoubleByte(uint16 chr) {
+bool GfxFontKorean::isDoubleByte(uint32 chr) {
 	uint16 ch = chr & 0xFF;
 	if ((ch >= 0xA1) && (ch <= 0xFE))
 		return true;
@@ -62,20 +62,20 @@ byte GfxFontKorean::getHeight() {
 		return _commonFont->getFontHeight() >> 1;
 }
 
-byte GfxFontKorean::getCharWidth(uint16 chr) {
+byte GfxFontKorean::getCharWidth(uint32 chr) {
 	if (getSciVersion() >= SCI_VERSION_2)
 		return _commonFont->getCharWidth(chr);
 	else
 		return _commonFont->getCharWidth(chr) >> 1;
 }
 
-void GfxFontKorean::draw(uint16 chr, int16 top, int16 left, byte color, bool greyedOutput) {
+void GfxFontKorean::draw(uint32 chr, int16 top, int16 left, byte color, bool greyedOutput) {
 	// TODO: Check, if character fits on screen - if it doesn't we need to skip it
 	_screen->putHangulChar(_commonFont, left & 0xFFC, top, chr, color);
 }
 
 #ifdef ENABLE_SCI32
-void GfxFontKorean::drawToBuffer(uint16 chr, int16 top, int16 left, byte color, bool greyedOutput, byte *buffer, int16 bufWidth, int16 bufHeight) {
+void GfxFontKorean::drawToBuffer(uint32 chr, int16 top, int16 left, byte color, bool greyedOutput, byte *buffer, int16 bufWidth, int16 bufHeight) {
 	byte *displayPtr = buffer + top * bufWidth + left;
 	// we don't use outline, so color 0 is actually not used
 	_commonFont->drawChar(displayPtr, chr, bufWidth, 1, color, 0, bufWidth - left, bufHeight - top);
